// pages/types.ts
export interface Expense {
  id: string;
  date: string;
  category: string;
  amount: number;
  payee: string;
  description?: string;
  user: string;
}